requirements.txt
datetime
Django==4.2.3
django-storages==1.13.2
boto3==1.26.157
botocore==1.29.157
docutils==0.19
gunicorn==20.1.0
jmespath==1.0.1
minio
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
psycopg2-binary==2.9.6
python-dotenv==1.0.1
PyYaml
qrcode
Pillow
APScheduler